fix(edgerun-rt): fix critical correctness and resource leak bugs
- reactor: switch from EPOLLET to level-triggered epoll to prevent hangs
  when tasks re-poll after EAGAIN (edge-triggered silently drops events)
- notify: eliminate race between atomic notify_one counter and waiters
  queue by unifying all state under single Mutex; fix waker index
  overwrite bug by removing index tracking entirely
- timers: fix CtrlC timer leak — register timer once and only
  re-register when deadline expires, not on every poll
- cancellation: fix waker Vec unbounded growth — register waker on
  first poll only instead of pushing on every re-poll
- io_traits: fix read_to_end returning partial data on Pending —
  stay pending until EOF instead of returning Ready prematurely
- reactor_fd_ready: remove unnecessary as i16 casts on POLLIN/POLLOUT
- trace: replace redundant closure in LazyLock::new
- notify: add impl Default, remove drop_non_drop in test

+++ crates/edgerun-rt/src/reactor_fd_ready.rs
@@ -46,7 +46,7 @@ impl Future for FdReadReady {
             revents: 0,
         };
         let res = unsafe { libc::poll(&mut pfd, 1, 0) };
-        if res > 0 && (pfd.revents & libc::POLLIN as i16) != 0 {
+        if res > 0 && (pfd.revents & libc::POLLIN) != 0 {
             self.done = true;
             return Poll::Ready(());
         }
@@ -87,7 +87,7 @@ impl Future for FdWriteReady {
             revents: 0,
         };
         let res = unsafe { libc::poll(&mut pfd, 1, 0) };
-        if res > 0 && (pfd.revents & libc::POLLOUT as i16) != 0 {
+        if res > 0 && (pfd.revents & libc::POLLOUT) != 0 {
             self.done = true;
             return Poll::Ready(());
         }
